AGRIChatBot/chatbot/requirements.txt
Flask==3.0.1
google-generativeai==0.4.0
googletrans==4.0.0-rc1

